MAJOR COMMIT! Partially modified code to expect gzipped files, and some modules will retroactively gzip uncompressed files in place during its first Output check (expect delays when re-running an older study).
Fixed bug in labels due to ImageMath "Byte" command (unwanted normalization); replaced with FSL command.

Finished forward operations of Kompactor (need to work Dekompressor code now!).  Added code for retroactively
masking warps to save significant space.  Currently dilating masks with radius = 5 voxels.  It has not been
completely proven that this is enough to avoid edge artifacts once a masked warp is applied.

Added safety check in case a slurm-out file is not produced.  It appears that when this happens, work will
appear to continue as normal, but the outputs will not be saved.  With code update, a slurmout-less job
will be cancelled and an email will be sent to the user...currently hardcoded as me (BJ -- rja20 @duke.edu).
Other jobs will continue to run.  This is the first implementation of using the email from cluster feature.

Undoubtedly there are other changes not noted here, due to the fact that this commit is a bit overdue.