This commit represents tweaks to convert_all_to_nifti and set_reference_space (and maybe some other minor changes).
convert_all_to_nifti:
This module is unique in that there is not a consistent module preceding it. Normally, the module before would check to make sure that
it had completed its work and therefore the files needed as input for the current module exist and are in place.  Until we implement a
proper data pulling module, we have to be aware of this exceptional behavior.  Before this commit, if work on a certain runno/contrast was
queued up to be performed, it would try to pull the input file from the ["pristine"] inputs directory without a proper check.  If the file
did not exist there, then the error message would be propagated as the file name to the matlab function.  Of course it would fail, but the
outputs were rather ugly to look at (in the log, and the .m files).

It is a bit tricky trying to figure out how to handle this, since the raw inputs are only needed for preprocessing/base_images purposes.
If these two stages are already completed and the results are in place, then it is irrelevant whether or not the raw inputs are in their
proper place.  In fact, it is often a disadvantage to unneccessarily haul the raw inputs around.  The upshot is that an init_check for the
raw inputs would be the ideal situation, except that that would need to include a preprocessing or base_images check to see which needed files
existed in the first place.  I may be able to make this happen by having the init_check call the output_check, compile a list of files that
it expects to process, checks to make sure those input files exists, then returns the init error message that some necessary files are missing.
The drawback to this is that often times we need the init_check to otherwise succeed in order for the inputs/work/results folder structure will
be created, giving as a location in which to put the input images in the first place.

set_reference_space:
There is some behaviour that is not thoroughly tested in regards to handling the rigid atlas.  This is partly due to trying to avoid the need
to specify whether or not a file is gzipped. I tried to fix it such that it would consistently check for the external atlas and determine if
it needed to be recentered around its center of mass. The standardized result is a gzipped file, and if the file exists but is not gzipped,
then the gzip command will be called.  Another aspect of this is how the file/filename is handled in the headfile, since it is expected to be
re-referenced as a version of it moves to base_images.  I think that I fixed things, but it is very possible that I introduced bugs or that
there were scenarios that I did not foresee which would break the updated code.  The context of this fix was that I had moved only essential
data off of glusterspace for a relatively outdated O'Brien ex vivo run, and was attempting to restore it to glusterspace then run the pipeline
(to obtain corrected jacobians). I.e., it was an attempt to make the pipeline more backwards compatible. If it breaks in the future, we should
look to this commit to see where I went awry.