This commit fixes at least two major issues that had been causing confusion: 1) surfstat_for_vbm_pipeline.m : Previously we were attempting to swap the sign when using 'jac' as the contrast.  Initially, that fix did exactly jac squat, as the default contrast that was being called was 'jac_from_MDT', so the condition for swapping sign was never triggered. ...until I modified the pipeline code to allow for specifying only those contrasts you wanted to be processed in VBA, in which case I would call 'jac'. ...until mid-December, when I realized that there should be no sign change, so all contrasts should be handled the same and no special handling was needed for jacobians. ...which was futile, as I corrected the right function but in the wrong folder.  The one I had modified was in civm_matlab_common_utils, but it was the one in vbm_pipeline (here) that was being called. ...SO, when I asked specifically for 'jac', it would finally trigger the condition to swap signs! Even though I thought that there were no conditions which would cause a swap.
So, beware of any jacobian results before this point, though they only ones that would be otherwise valid (see point \#2 below), are the McNamara
ones that I used for the paper and poster--those jacobians are actually right but th jac VBA have the sign swapped (so Control_gt_KA is actually
KA_gt_Control,etc). In the various SyN folders, these correspond to faMDT_Control_n10 (KA) and faMDT_Control_n10d (Phantom)--b and c should be
ignored as they are bad for a variety of reasons.

2) calculate_jacobian.m, jacobian_2.m, and calculate_jacobians_vbm.pm :
Whew! What a mess this has been. The ANTs function for calculating jacobians is completely unreliable (and perhaps even incapable) when
it comes to producing accurate results.  Until the get a proper fix in place (and we get it on the cluster), then we need to use the
(incredibly slow) stopgap of my matlab script that will at least get it as right as finite-differences can.  I suspect that we really
want the capability to do it geometrically, as we know we are currently biased towards detecting atrophy.

This implemenation CANNOT handle arbitrary [header] orientations, and only works for standard CIVM headers that 1) centered about the origin, and 2) have (-1 -1 1)
diagonal for the rotation matrix, and the corresponding signage in the s_row fields.  Requirement 2 is a must...requirement 1 may work otherwise
but is untested.

3) Unsure if there are other updates to code represented by this commit...I really need to commit more regularly.