First commit of ANTsR code...though there still is a sign flip bug.  I hope to find the culprit within the very near future. I believe it lies in the "diagnosis  variable in ANTsR_vba_fx.R/ANTsR_vba.R, and how it is interpreted in simple_voxel_based_analysis.R.
MORE IMPORTANTLY, we finally fixed a devious bug lurking in SurfStat.  In addition to the many sign flip errors/cancellation of errors that we have been battling, this bug would reassign each image to the wrong group...usually.  That was the case for equal numbers of control and treated images, and the result would be valid except for a sign flip ("x greater than y" would actually represent "y greater than x", and vice versa).  If the group sizes were unequal, the population images of each group would change according to the difference in size between the two...and the results should be considered completely and utterly useless.