VBA_pipeline in definite state of non-equilibrium.  Major recent changes addressed the so-called "recenter while setting reference space" which accidentally created a translation for each runno/contrast combination, due to the fact that "masked" was being caught as a contrast, thus the remainder (runno AND actual contrast) was being considered a unique runno (the translations are supposed to be created for each runno, which it is assumed to have all contrasts co-registered).
Beyond that, there is some know, yet unresolved issues with the how the rigid atlas is handled during the preprocess/base_images stages.  This seems to center around the .gz extension (and therefore whether or not it has been compressed).

THIS IS THE LAST COMMIT BEFORE MAKING MAJOR CHANGES TO ACCOMODATE SyGN ITERATIVE TEMPLATE CREATION!!!