First semi-functional end-to-end usable version of single_segmentation.
This includes pulling tensor_create results data from a specific recon machine, and calculating a combine bval/bvec b-table from the headefile in the case that tensor_create work data is no longer there (this was the case with my test data, otherwise I probably would not have gone this route).  This means that everything that is needed can be pulled over in one shot by transferring the whole tensor-results folder.  On the other end, the bvecs are transformed, and if we decide to convert to RAS, are copied into the results directory.  This is still imperfect, and doesn't do a good job of file checking and then acting appropriately.

Code still needs to be cleaned up, and better handling of converting to RAS needs to be implemented as well.  Also should be noted that there needs to be a concurrent update with civm_matlab_utilities folder (or whatever it is exactly called).  Unsure if any there are any dependencies on recent changes to pipeline_utilities.