Write_individual_label_stats has been integrated into pipeline.
It appears to pass most tests.  It successfully checks existing files to see what contrasts have already been completed, and will not schedule a job if it it is not necessary.  The matlab executable, for its part, will not redo work for a particular contrast which has already been completed (thanks to an internal check).  The fprintf output will say "processing contrast n of N" or something like that, where the total N will include all contrasts, and doesn't necessarily mean that that many are being currently [re-]processed.