Pre-processing for non-parametric testing incorporated.
Compared to previous commit, the strategy for achieving NPT has changed, moving from stand-alone modules to simply an optional "software" for vbm analysis.  In some aspects, this is not the best approach, since there is more branches of work to be performed; each of these stages shouls have their own work checking, etc.  But alas, so much of the work is duplicated (namely, smoothing the images--which we want, right?), so it makes sense that it's just another form of VBA.

ALSO, we tracked down a "get_value()" in the Init_check that should have been "$Hf->get_value()" that was causing an early exit from the Init_check, and thus not properly catching any bad requests for VBA software.  A quick audit of the code appeared to report that this was the only instance of an unintentionally bare et_value/set_value, so hopefully this is the only place this ugly bug raises its head.


Former-commit-id: 8520f5436d0bbb8342a2048e2d73f65e57b0e1d5