Non-parametric testing almost finished...maybe 80% done. Also other minor tweaks.
The minor changes include handling output from cluster_exec, which now supports double_booking--er, I mean backup slurm calls. The output when this feature is on is one string of two job names joined by a comma (the second job is the backup to the first).  Also, tried to change all non-job calls to return zero, the same as failed job submission. That way I can simply test to see if $job is not zero/null, instead of forcing it to be handled as a string or numeric.

Also, I'm about to try to remove a previously committed nii.gz file, which is preventing the pushing of the code to github.