Medium code update.
Added matlab executable version of strip_mask.  Changed reorientation such that it will use the matlab executable that is at the end of the pipeline for RAS'ing results.

Moved pull_civm_tensor_data's call such that it will only run if there missing files in the -work/preprocess folder.

Also, this is an overdue update that represents many smaller tweaks to the code which I can't exhaustively recall right now.