Adding some SGE support files, preparing to add isotropic enforcement.
Other fixes include how the preprocess files are normalized, with or without "masked" suffix. This relates to the code trying to determine a specimen/subject of median volume to be used as an affine target, code was touched up there as well.

Some paths to matlab_execs were updated.

force_isotropic_resolution variable added, though this feature is coming after this commit, and can be ignored for now.

Fixed the command used for creating identity_warp.nii.gz.