Stats Calc and cluster_exec copy-pasta update
Fixed up stats calcs and their calls to be more robust, specifically
regarding the files looked for and when we look.
Many examples of nearly the same chain of lines in prep to run cluster_exec have been made as
similar as possible, highlighting a great deal of "there could be a function for that".

Added channel spec to stats main call because we sometimes intentionally omit, or adjust
channels before we run through.

Made qa cov calc resilient to missing contrasts so it'll finish the job as well as it can.