Fixed bugs in transform capture,
changed internal nomenclature for "road_(forward|backward)"
It now is always talking about the current package, with the road
forward going to the target, and backward being the road coming
from the target.
Formerly it might have been trying to talk about the path of labels,
but that is invisible inside the code making it hard to keep straight.