Spaces not escaped with a backslash in the headfile will be removed.
This should help prevent typos from derailing a SAMBA run.

Spaces that ARE escaped with a backslash will immediately have the backslashes removed upon initial processing.

It is thought that at some point in time a user will need to include a path that (ill-advisedly) has spaces in it.  HOWEVER, given that SAMBA operates in both PERL and Bash, it is almost inevitable that paths-with-spaces will break the pipeline in multiple places. NO FURTHER SUPPORT FOR THESE PATH WILL BE PROVIDED.

There is a very good reason to not have spaces in your paths and filenames, and so it is highly recommended to adopt the practice of avoiding them.